\name{LSDeaths}
\docType{data}
\alias{LSDeaths}
\title{ONS Longitudinal Study Mortality Data}
\description{
Some mortality data from the UK Office for National Statistics Longitudinal Study (LS), taken from Sergeant and Firth (2004).  The data pertain to males in the LS aged 25-64 at the 1991 UK census who were still alive and present on 1 January 1996.  The number of deaths in the period 1 January 1996 to 31 December 2000 and the number of individuals at risk at the start of the period (i.e. the number that died plus the number that survived) are cross-classified by social class and age, both recorded at the 1991 census.  Age is given in ten-year intervals and the Registrar General's social classification is used, which groups individuals in a graded hierarchy of occupations according to skill as follows:

\tabular{ll}{
Social class \tab Description\cr
V \tab Unskilled Occupations\cr
IV \tab Partly-Skilled Occupations\cr
IIIM \tab Skilled Manual Occupations\cr
IIIN \tab Skilled Non-manual Occupations\cr
II \tab Managerial and Technical Occupations\cr
I \tab Professional Occupations
}
}
\usage{data(LSDeaths)}
\format{A data frame with 24 rows.}
\references{ Sergeant, J. C. and Firth D. (2004)  Relative index of inequality: definition, estimation and inference.  In preparation }
\keyword{datasets}
