% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_items.R
\name{print.likert_item_scan}
\alias{print.likert_item_scan}
\title{Print a likert_item_scan}
\usage{
\method{print}{likert_item_scan}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"likert_item_scan"}.}

\item{...}{Further arguments (currently unused).}
}
\description{
Print a likert_item_scan
}
