% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_likert.R
\name{print.likert_text_scan}
\alias{print.likert_text_scan}
\title{Print a likert_text_scan}
\usage{
\method{print}{likert_text_scan}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"likert_text_scan"}.}

\item{...}{Further arguments (currently unused).}
}
\description{
Print a likert_text_scan
}
