% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.rii_table}
\alias{print.rii_table}
\title{Print an rii_table}
\usage{
\method{print}{rii_table}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"rii_table"}.}

\item{...}{Further arguments (currently unused).}
}
\description{
Print an rii_table
}
